+++ receiver/hostmetricsreceiver/internal/scraper/cpuscraper/cpu_scraper.go
@@ -57,7 +57,7 @@ func (s *scraper) start(context.Context, component.Host) error {
 func (s *scraper) scrape(_ context.Context) (pdata.MetricSlice, error) {
 	metrics := pdata.NewMetricSlice()
 
-	now := pdata.TimestampFromTime(time.Now())
+	now := pdata.NewTimestampFromTime(time.Now())
 	cpuTimes, err := s.times( /*percpu=*/ true)
 	if err != nil {
 		return metrics, scrapererror.NewPartialScrapeError(err, metricsLen)