+++ receiver/hostmetricsreceiver/internal/scraper/loadscraper/load_scraper.go
@@ -57,7 +57,7 @@ func (s *scraper) shutdown(ctx context.Context) error {
 func (s *scraper) scrape(_ context.Context) (pdata.MetricSlice, error) {
 	metrics := pdata.NewMetricSlice()
 
-	now := pdata.TimestampFromTime(time.Now())
+	now := pdata.NewTimestampFromTime(time.Now())
 	avgLoadValues, err := s.load()
 	if err != nil {
 		return metrics, scrapererror.NewPartialScrapeError(err, metricsLen)