+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Remove deprecated `tracetranslator.AttributeValueToString` and `tracetranslator.AttributeMapToMap` (#3873)
 - Change semantic conventions for status (code, msg) as per specifications (#3872)
+- Add `pdata.NewTimestampFromTime`, deprecate `pdata.TimestampFromTime` (#3868)
 
 ## v0.33.0 Beta